More source-related cleanup and cruft removal.
* Move baseVCSSource into vcs_source.go
* Rename source_test.go to vcs_source_test.go
* Remove sourceExistence type and remaining uses of it
