+++ solve_basic_test.go
@@ -1405,10 +1405,6 @@ func (sm *depspecSourceManager) GetManifestAndLock(id ProjectIdentifier, v Versi
 	return nil, nil, fmt.Errorf("Project %s at version %s could not be found", id.errString(), v)
 }
 
-func (sm *depspecSourceManager) AnalyzerInfo() (string, int) {
-	return "depspec-sm-builtin", 1
-}
-
 func (sm *depspecSourceManager) ExternalReach(id ProjectIdentifier, v Version) (map[string][]string, error) {
 	pid := pident{n: ProjectRoot(id.normalizedSource()), v: v}
 	if m, exists := sm.rm[pid]; exists {
